Add the notion of a tree descriptor
There is a lot of important information provided to us via the addon/app models that make things like path resolution and parent child relationship easier to reason about.  This adds the idea of building up tree descriptors and also tagging trees with names as they go through the JS processing. This information can then be used down stream in the pre-packager.