Introduce module to set the load path of addons
This introduces a "load path" broccoli plugin that sets the load path for addons.  It supports the existing addon structure which looks like the following:

my-cool-addon/
  addon/
    helpers/
      some-cool-helper.js
    index.js

While this works and how every single addon is writen today, we can do better by providing multiple loadpaths into an addon directory. What this means that a single addon can have multiple require paths. This requires a new addon structure if you have a need for this. The directory structure is as follows

my-cool-addon/
  addon/
    my-cool-addon/
      helpers/
        some-helper.js
    my-other-cool-thing/
      mixins/
        blamo.js
    my-cool-addon.js

The files in the root of the addon directory are "main" files for the addon and traditionally reexport files.