Assembler refactor
Prior to this commit it was extremely non-obvious on how metadata was
schelped around. This commit introduces the concept of a tree descriptor
model which holds meta data about packages that eventually manifest
themselves in trees.

There is still the concept of holding the trees seperate as long as
possible, but now we have better ways of accessing those trees when we
need them.